Restructure of methods.py and top.py. Broke them out into classes in ./methods/ and ./top/
