+++ dyc/top/TopFormatter.py
@@ -4,79 +4,13 @@ the application. Here is the the place where the actual reading,
 parsing and validation of the files happens.
 """
 import sys
-import re
-import fileinput
 import copy
-import linecache
 import click
-from .utils import (
-    get_leading_whitespace,
+from ..utils import (
     BlankFormatter,
     get_indent,
     add_start_end,
-    is_one_line_method, ##TODO remove; unneeded for file-level documentation
 )
-from .base import Builder
-import os
-
-
-class TopBuilder(Builder):
-    already_printed_filepaths = []
-
-    def extract_and_set_information(self, filename, start, line, length):
-        """
-        This is a main abstract method tin the builder base
-        to add result into details. Used in Top Builder to
-        pull the candidates that are subject to docstrings
-        Parameters
-        ----------
-        str filename: The file's name
-        int start: Starting line
-        str line: Full line text
-        int length: The length of the extracted data
-        """
-        start_line = linecache.getline(filename, start)
-        initial_line = line
-        start_leading_space = get_leading_whitespace(
-            start_line
-        )  # Where function started
-        top_string = start_line
-        if not is_one_line_method(start_line, self.config.get("keywords")):
-            top_string = line
-            linesBackwards = top_string.count("\n") - 1
-            start_leading_space = get_leading_whitespace(
-                linecache.getline(filename, start - linesBackwards)
-            )
-        line_within_scope = True
-        lineno = start + 1
-        line = linecache.getline(filename, lineno)
-        end_of_file = False
-        end = None
-        while line_within_scope and not end_of_file:
-            current_leading_space = get_leading_whitespace(line)
-            if len(current_leading_space) <= len(start_leading_space) and line.strip():
-                end = lineno - 1
-                break
-            top_string += line
-            lineno = lineno + 1
-            line = linecache.getline(filename, int(lineno))
-            end_of_file = True if lineno > length else False
-
-        if not end:
-            end = length
-
-        linecache.clearcache()
-        return TopInterface(
-            plain=top_string,
-            name=self._get_name(initial_line),
-            start=start,
-            end=end,
-            filename=filename,
-            arguments=self.extract_arguments(initial_line.strip("\n")),
-            config=self.config,
-            leading_space=get_leading_whitespace(initial_line),
-            placeholders=self.placeholders,
-        )
 
 
 class TopFormatter:
@@ -168,12 +102,12 @@ class TopFormatter:
         top_split = self.plain.split("\n")
         if self.config.get("within_scope"):
             # Check if method comes in an unusual format
-            keywords = self.config.get("keywords")
-            firstLine = top_split[0]
+            # keywords = self.config.get("keywords")
+            # firstLine = top_split[0]
             pos = 1
-            while not is_one_line_method(firstLine, keywords):
-                firstLine += top_split[pos]
-                pos += 1
+            # while not is_one_line_method(firstLine, keywords):
+            #     firstLine += top_split[pos]
+            #     pos += 1
             top_split.insert(pos, polished)
         else:
             top_split.insert(0, polished)
@@ -214,49 +148,3 @@ class TopFormatter:
             self.result = polished
         else:
             self.confirm(polished)
-
-
-class TopInterface(TopFormatter):
-    def __init__(
-        self,
-        plain,
-        name,
-        start,
-        end,
-        filename,
-        arguments,
-        config,
-        leading_space,
-        placeholders,
-    ):
-        self.plain = plain
-        self.name = name
-        self.start = start
-        self.end = end
-        self.filename = filename
-        self.arguments = arguments
-        self.top_docstring = ""
-        self.config = config
-        self.leading_space = leading_space
-        self.placeholders = placeholders
-
-    def prompt(self):
-        """
-        Wrapper method for prompts and calls for prompting
-        file then formats them
-        """
-        self._prompt_docstring()
-        self.format()
-
-    def _prompt_docstring(self):
-        """
-        Simple prompt for a file's docstring
-        """
-        if self.placeholders:
-            self.top_docstring = "<docstring>"
-        else:
-            echo_name = click.style(self.name, fg="green")
-            self.top_docstring = click.prompt(
-                "\n({}) Top docstring ".format(echo_name)
-            )
-        pass
